QKV and MLP matmul combination for LLaMA (#543)
This PR supports QKV and MLP matmul combination for LLaMA.

It adds an option to LlamaConfig, which controls if to combine the QKV
computation. Right now, we will combine the QKV and MLP matmul as long
as the quantization method is not AutoGPTQ. The option can be also
manually set in llama.py.